requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
spotipy
python-dotenv
sqlalchemy
psycopg2-binary